+++ scripts/test/Selenium/Agent/Admin/AdminSystemConfigurationComplex.t
@@ -12,7 +12,6 @@ use utf8;
 
 use vars (qw($Self));
 
-# get selenium object
 my $Selenium = $Kernel::OM->Get('Kernel::System::UnitTest::Selenium');
 
 my @Tests = (
@@ -379,6 +378,101 @@ my @Tests = (
             {
                 '3th' => '3',
                 '4th' => '4',
+            },
+        ],
+    },
+    {
+        Name     => 'ExampleAoHEmpty',
+        Commands => [
+            {
+                Hover => '.Content',
+            },
+            {
+                JqueryClick => '.SettingEdit',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .AddArrayItem',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .AddArrayItem',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .AddArrayItem',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .ArrayItem:nth-of-type(2) > .RemoveButton',
+            },
+            {
+                Select => '.Setting > .Array > .ArrayItem:nth-of-type(1) .HashItem:nth-of-type(1) input',
+            },
+            {
+                # Key
+                Write => '1st',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .ArrayItem:nth-of-type(1) .HashItem:nth-of-type(1) .AddKey',
+            },
+            {
+                Select =>
+                    '.Setting > .Array > .ArrayItem:nth-of-type(1) .HashItem:nth-of-type(1) .SettingContent input',
+            },
+            {
+                # Value
+                Write => '1',
+            },
+            {
+                Select => '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(1) input',
+            },
+            {
+                # Key
+                Write => '2nd',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(1) .AddKey',
+            },
+            {
+                Select =>
+                    '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(1) .SettingContent input',
+            },
+            {
+                # Value
+                Write => '2',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .ArrayItem:nth-of-type(2) .AddHashKey',
+            },
+            {
+                Select => '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(2) input',
+            },
+            {
+                # Key
+                Write => '3th',
+            },
+            {
+                JqueryClick => '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(2) .AddKey',
+            },
+            {
+                Select =>
+                    '.Setting > .Array > .ArrayItem:nth-of-type(2) .HashItem:nth-of-type(2) .SettingContent input',
+            },
+            {
+                # Value
+                Write => '3',
+            },
+            {
+                JqueryClick => '.Update',
+            },
+            {
+                Select => 'input',
+            },
+        ],
+        ExpectedResult => [
+            {
+                '1st' => '1',
+            },
+            {
+                '2nd' => '2',
+                '3th' => '3',
             },
         ],
     },
@@ -1418,7 +1512,6 @@ my @Tests = (
 $Selenium->RunTest(
     sub {
 
-        # get helper object
         my $Helper          = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
         my $SysConfigObject = $Kernel::OM->Get('Kernel::System::SysConfig');